Notes and Users
Updated the registration page and  the controller action with data validation and notifications.
Updated the login page and the controller action with data validation and notifications.
Extracted user validation constants.
Updated the User data model.
Notes can now be added and viewed trough the request details page, in addition to the requests table,